+++ packages/components/src/Form/Inputs/Combobox/utils/parseOption.ts
@@ -24,6 +24,17 @@
 
  */
 
-export * from './getComboboxText'
-export * from './formatOptionAsString'
-export * from './parseOption'
+import { ComboboxOptionObject } from '../types'
+
+export function parseOption(text: string) {
+  try {
+    const parsed = JSON.parse(text)
+    if (parsed.value) {
+      return parsed as ComboboxOptionObject
+    } else {
+      return { value: text }
+    }
+  } catch (e) {
+    return { value: text }
+  }
+}